[Feature] 로그인 기능 (#23)
* 사용자 권한 추가
* 공통 응답 DTO, 커스텀 Exception, ExceptionHandler 추가
* 입력값 검증을 위해 @Validation 의존성을 추가함.
* 'MethodArgumentNotValidException' 핸들러 메서드 추가
* context-path '/api/'로 설정
* 시큐리티(인증, 인가) 기능 구현
* 유저 MOCK 데이터 추가
* 개발 단계에서 사용할 로깅 레벨 설정
* 인증/인가 테스트 용 API 추가
* 인증 구현에 따른 'JPA AuditorAware' 구현
* #5 - 'BoardUserDetailsService' 단위 테스트

+++ src/main/java/com/back/repository/HashtagRepository.java
@@ -3,5 +3,5 @@ package com.back.repository;
 import com.back.domain.Hashtag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HashtagRepository extends JpaRepository<Hashtag, Integer> {
+public interface HashtagRepository extends JpaRepository<Hashtag, Long> {
 }
